Adds .csv reader, writer and parser
The .csv file is read, the header is stripped and the data is saved in
memory in a std::map where the key is the DBN and the value is a struct
with all the fields. The std::map was chosen to index the DBN and
optimize searches.

To validate reading the file, parsing to a data structure, parsing the
struct to a string and writing the result in a file, a routine was added
in main and will be removed when the CLI interface is implemented.

Although the specs say the DBN is unique, the provided .csv have two
entries with the same DBN, but different data (04M610 and 10X430). The
application keeps the first entry read from the file and discards other
entries with the same DBN.

The application isn't yet able to create the directory `output`. OS
agnostic folder creation is problematic before C++14/C++17. I'm not sure
if it's better to bump the C++ standard version or implement the POSIX
directory creation.