Adds interactive CLI with search by name & DBN
An interactive CLI was created, implementing search by DBN, search by
School Name and saving results of Name searches to a file.

Case insensitive School name searches are not implemented yet.

One bug was found in the search by School Name. Search results are being
saved in a std::map for School Name ordering and to reuse methods
created before. But School Names are not unique, and name collisions
cause the search data to be garbled. The name search must be refactored
to use different structures to hold data and to order search results.